Move utils tests to test_utils.py
