+++ tests/genai/discovery/test_pipeline.py
@@ -13,7 +13,6 @@ from mlflow.genai.discovery.pipeline import (
     discover_issues,
 )
 from mlflow.genai.discovery.utils import (
-    format_trace_content,
     verify_scorer,
 )
 from mlflow.genai.evaluation.entities import EvaluationResult
@@ -606,19 +605,6 @@ def test_annotate_traces_session_level_logs_on_first_trace():
     assert session_ids == {"session-A", "session-B"}
 
 
-def testformat_trace_content_includes_errors(make_trace):
-    trace = make_trace(error_span=True)
-    content = format_trace_content(trace)
-    assert "Errors:" in content
-    assert "Connection failed" in content
-
-
-def testformat_trace_content_no_errors(make_trace):
-    trace = make_trace()
-    content = format_trace_content(trace)
-    assert "Errors:" not in content
-
-
 def test_recluster_merges_similar_singletons():
     analyses = [
         _ConversationAnalysis(
@@ -889,49 +875,3 @@ def test_build_satisfaction_instructions_no_categories(categories):
 
     instructions = build_satisfaction_instructions(use_conversation=True, categories=categories)
     assert "issue categories" not in instructions
-
-
-def test_collect_example_trace_ids_gathers_from_analyses():
-    from mlflow.genai.discovery.utils import collect_example_trace_ids
-
-    analyses = [
-        _ConversationAnalysis(
-            full_rationale="r1", affected_trace_ids=["t1", "t2"], execution_path="p1"
-        ),
-        _ConversationAnalysis(full_rationale="r2", affected_trace_ids=["t3"], execution_path="p2"),
-    ]
-    issue = _IdentifiedIssue(
-        name="test", description="d", root_cause="rc", severity="high", example_indices=[0, 1]
-    )
-    result = collect_example_trace_ids(issue, analyses)
-    assert result == ["t1", "t2", "t3"]
-
-
-def test_collect_example_trace_ids_skips_out_of_bounds():
-    from mlflow.genai.discovery.utils import collect_example_trace_ids
-
-    analyses = [
-        _ConversationAnalysis(full_rationale="r1", affected_trace_ids=["t1"], execution_path="p1"),
-    ]
-    issue = _IdentifiedIssue(
-        name="test", description="d", root_cause="rc", severity="high", example_indices=[0, 5]
-    )
-    result = collect_example_trace_ids(issue, analyses)
-    assert result == ["t1"]
-
-
-def test_collect_example_trace_ids_caps_at_max(monkeypatch):
-    from mlflow.genai.discovery import utils
-    from mlflow.genai.discovery.utils import collect_example_trace_ids
-
-    monkeypatch.setattr(utils, "MAX_EXAMPLE_TRACE_IDS", 2)
-    analyses = [
-        _ConversationAnalysis(
-            full_rationale="r1", affected_trace_ids=["t1", "t2", "t3"], execution_path="p1"
-        ),
-    ]
-    issue = _IdentifiedIssue(
-        name="test", description="d", root_cause="rc", severity="high", example_indices=[0]
-    )
-    result = collect_example_trace_ids(issue, analyses)
-    assert len(result) == 2
